pruning unnecessary dependencies 2
