+++ client/Cargo.toml
@@ -17,14 +17,13 @@ edition = "2021"
 maintenance = { status = "actively-developed" }
 
 [features]
-wbindgen = [ "naia-socket-shared/wbindgen", "wasm-bindgen", "js-sys", "web_sys", "serde", "serde_derive", "serde_json", "regex" ]
-mquad = [ "naia-socket-shared/mquad", "miniquad", "regex" ]
+wbindgen = [ "naia-socket-shared/wbindgen", "wasm-bindgen", "js-sys", "web_sys", "serde", "serde_derive", "serde_json" ]
+mquad = [ "naia-socket-shared/mquad", "miniquad" ]
 
 [dependencies]
 naia-socket-shared = { path = "../shared", version = "=0.8.0" }
 cfg-if = "0.1.10"
 log = { version = "0.4" }
-url = { version = "2.2.2" }
 wasm-bindgen = { version = "0.2.68", optional = true  }
 js-sys = { version = "0.3", optional = true  }
 web_sys = { version = "0.3.22", package = "web-sys", features = [
@@ -36,4 +35,9 @@ serde = { version = "^1.0.59", optional = true  }
 serde_derive = { version = "^1.0.59", optional = true  }
 serde_json = { version = "^1.0.59", optional = true }
 miniquad = { version = "=0.3.0-alpha.28", features = ["log-impl"], optional = true }
-regex = { version = "1.5.4", optional = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+local_ipaddress = { version = "0.1.3" }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+regex = { version = "1.5.4" }
