+++ client/src/impls/native/socket.rs
@@ -7,7 +7,7 @@ use std::{
 
 use log::info;
 
-use naia_socket_shared::{find_my_ip_address, parse_server_url, url_to_socket_addr, SocketConfig};
+use naia_socket_shared::{parse_server_url, url_to_socket_addr, SocketConfig};
 
 use crate::{
     io::Io,
@@ -94,5 +94,20 @@ impl Socket {
             .expect("Socket is not connected yet! Call Socket.connect() before this.")
             .packet_receiver
             .clone();
+    }
+}
+
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
+
+/// Helper method to find local IP address, if possible
+pub fn find_my_ip_address() -> Option<IpAddr> {
+    let ip = local_ipaddress::get().unwrap_or_default();
+
+    if let Ok(addr) = ip.parse::<Ipv4Addr>() {
+        return Some(IpAddr::V4(addr));
+    } else if let Ok(addr) = ip.parse::<Ipv6Addr>() {
+        return Some(IpAddr::V6(addr));
+    } else {
+        return None;
     }
 }
