+++ server/Cargo.toml
@@ -21,7 +21,7 @@ features = [ "use-webrtc" ]
 
 [features]
 use-udp = [ "async-io" ]
-use-webrtc = [ "webrtc-unreliable", "async-dup", "http", "futures-core" ]
+use-webrtc = [ "webrtc-unreliable", "async-dup", "futures-core", "http" ]
 
 [dependencies]
 naia-socket-shared = { path = "../shared", version = "=0.8.0" }
@@ -32,9 +32,8 @@ futures-util = { version = "0.3", features = ["sink"] }
 smol = { version = "1.2.4" }
 crossbeam = { version = "0.8" }
 once_cell = { version = "1.4.1" }
-futures-lite = { version = "1.11.0" }
 async-io = { version = "1.1", optional = true }
-futures-core = { version = "0.3", optional = true }
 webrtc-unreliable = { version = "0.5.1", optional = true }
 async-dup = { version = "1.2.2", optional = true }
+futures-core = { version = "0.3", optional = true }
 http = { version = "0.2", optional = true }
