+++ shared/src/lib.rs
@@ -16,14 +16,12 @@ extern crate cfg_if;
 /// conditions
 pub mod link_condition_logic;
 
-mod find_my_ip_address;
 mod impls;
 mod link_conditioner_config;
 mod socket_config;
 mod time_queue;
 mod url_parse;
 
-pub use find_my_ip_address::find_my_ip_address;
 pub use impls::{Instant, Random, Timer, Timestamp};
 pub use link_conditioner_config::LinkConditionerConfig;
 pub use socket_config::SocketConfig;
